Ran include-what-you-use on the codebase
This ensures that transitive includes are directly present in the files
where they are needed. Used --keep=* option to only add new headers and
forward declarations and not remove any in the process. Given that the
change only adds existing includes already present transitively, the
build time remains the same in the end.